style: 불필요한 폴더 삭제 및 login 객체 분리
